Update models/P3Page.php
P3Page::getMenuItems() now returns sub-menu-structure if a menu-item has children.

Notice that the 'url' key in the 'if' statement is set to '#' when it has children. Twitter Bootstrap needs the 'href' set to be '#' when it has to be a dropdownmenu. 
The implication is that the parent P3Page does not work as a hyperlink even when a route is set.